+++ member-service/src/main/java/com/green/member/client/CoreServiceClient.java
@@ -1,6 +1,6 @@
 package com.green.member.client;
 
-import com.green.common.model.GpaResult;
+import com.green.common.client.GpaResult;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
